Add support for `serverless.apigateway.API` (#6)
Adds a new API type to expose an APIGateway RestAPI with three primary features:

1. Automatically creates the RestAPI, Stage and Deployment needed to manage an API
2. Allows passing either a Swagger spec or a higher-level bag of route handlers for easy (but fully flexible) configuration.
3. Accepts Handlers as description of route callbacks, supporting either AWS Lambda objects or JavaScript callbacks.

Naming is still up for discussion (along with general expected renaming in this library). Also, we will likely want to add additional capabilities to this library so that it can be used in @pulumi/cloud-aws instead of duplicating functionality in both locations.